[1191] Support configuring project.name via the properties file
